% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mbhl_client_info}
\alias{mbhl_client_info}
\title{Display a summary of the client database state}
\usage{
mbhl_client_info(con)
}
\arguments{
\item{con}{A DBI connection or pool object.}
}
\value{
Invisibly returns a named list with the counts.
}
\description{
Display a summary of the client database state
}
