% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{mbhlcore_permissions}
\alias{mbhlcore_permissions}
\title{Registry of mbhlCore permissions}
\usage{
mbhlcore_permissions
}
\description{
Exported list of all permissions defined by this package, with human-readable
descriptions. Used by the admin interface to generate permission toggles
dynamically.
}
