% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrations.R
\name{migration_status}
\alias{migration_status}
\title{Display migration status in the console}
\usage{
migration_status(con)
}
\arguments{
\item{con}{A DBI connection or pool object.}
}
\value{
Invisibly returns the data.frame from \code{\link[=list_migrations]{list_migrations()}}.
}
\description{
Display migration status in the console
}
