% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_aircraft_types}
\alias{seed_demo_aircraft_types}
\title{Seed demo: types d'aéronefs}
\usage{
seed_demo_aircraft_types(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: types d'aéronefs
}
