% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_company_contacts}
\alias{seed_demo_company_contacts}
\title{Seed demo: contacts compagnies}
\usage{
seed_demo_company_contacts(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: contacts compagnies
}
