% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_exchange_rates}
\alias{seed_demo_exchange_rates}
\title{Seed demo: taux de change}
\usage{
seed_demo_exchange_rates(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: taux de change
}
