% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_part_types}
\alias{seed_demo_part_types}
\title{Seed demo: types de pièces}
\usage{
seed_demo_part_types(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: types de pièces
}
