% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_parts_catalog_applicability}
\alias{seed_demo_parts_catalog_applicability}
\title{Seed demo: applicabilité des pièces par type d'aéronef}
\usage{
seed_demo_parts_catalog_applicability(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: applicabilité des pièces par type d'aéronef
}
