% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_parts_catalog_positions}
\alias{seed_demo_parts_catalog_positions}
\title{Seed demo: positions d'installation des pièces}
\usage{
seed_demo_parts_catalog_positions(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: positions d'installation des pièces
}
