% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_personnel_licenses}
\alias{seed_demo_personnel_licenses}
\title{Seed demo: licences du personnel}
\usage{
seed_demo_personnel_licenses(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: licences du personnel
}
