% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_publication_applicability}
\alias{seed_demo_publication_applicability}
\title{Seed demo: applicabilité des publications}
\usage{
seed_demo_publication_applicability(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: applicabilité des publications
}
