% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_publication_revision_log}
\alias{seed_demo_publication_revision_log}
\title{Seed demo: journal de révisions des publications}
\usage{
seed_demo_publication_revision_log(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: journal de révisions des publications
}
