% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_demo.R
\name{seed_demo_publications}
\alias{seed_demo_publications}
\title{Seed demo: publications}
\usage{
seed_demo_publications(con)
}
\arguments{
\item{con}{Connexion ou pool DBI.}
}
\description{
Seed demo: publications
}
